combine test files, refactor
